restructure db
