+++ db/models/index.js
@@ -5,7 +5,7 @@ const Sequelize = require('sequelize');
 
 const basename = path.basename(__filename);
 const env = process.env.NODE_ENV || 'development';
-const config = require('../config').mysql[env];
+const config = require('../config')[env];
 
 const db = {};
