+++ db/seeders/20191203025042-insert-favorites.js
