+++ package.json
@@ -7,7 +7,7 @@
     "node": ">=6.13.0"
   },
   "scripts": {
-    "start": "node -r dotenv/config server/server.js",
+    "start": "node -r dotenv/config server.js",
     "dev": "webpack --mode development -d --watch",
     "build": "webpack -p",
     "deploy:s3": "npm run build && grunt aws_s3",
