+++ router/api/songs/get.js
@@ -1,5 +1,5 @@
 const { Op } = require('sequelize');
-const db = require('../../../models');
+const db = require('../../../db/models');
 
 const getSong = (req, res) => db.track
   .findOne({
