+++ server.js
@@ -1,4 +1,4 @@
-const app = require('../app');
+const app = require('./app');
 
-const PORT = 3004;
+const PORT = process.env.PORT || 3004;
 app.listen(PORT, () => console.log(`Server listening on port ${PORT}`));
